+++ src/components/pages/SignUp.js
@@ -1,21 +1,22 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import { useForm } from 'react-hook-form';
 import { useDispatch } from 'react-redux';
-import { signin } from '../redux/auth';
+import { signup } from '../../redux/auth';
 
-const SignIn = () => {
+const SignUp = () => {
   const dispatch = useDispatch();
   const { register, handleSubmit } = useForm();
 
-  const onSubmit = (data) => dispatch(signin(data));
+  const onSubmit = (data) => dispatch(signup(data));
 
   return (
     <form onSubmit={handleSubmit(onSubmit)} method="post">
+      <input placeholder="Enter your name" {...register('name', { required: true })} />
       <input type="email" placeholder="Enter your email" {...register('email', { required: true })} />
       <input type="password" placeholder="Enter your password" {...register('password', { required: true })} />
-      <input type="submit" value="Log In" />
+      <input type="submit" value="Create account" />
     </form>
   );
 };
 
-export default SignIn;
+export default SignUp;
